Remove getChannelClient methods. Rename ChannelClientImpl
