+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -19,8 +19,6 @@ import io.getstream.chat.android.client.api.models.UpdateChannelRequest
 import io.getstream.chat.android.client.call.Call
 import io.getstream.chat.android.client.call.map
 import io.getstream.chat.android.client.channel.ChannelClient
-import io.getstream.chat.android.client.controllers.ChannelController
-import io.getstream.chat.android.client.controllers.ChannelControllerImpl
 import io.getstream.chat.android.client.di.ChatModule
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.events.ChatEvent
@@ -716,41 +714,12 @@ public class ChatClient internal constructor(
     }
 
     /***
-     * Returns a [ChannelController] for given type and id
-     *
-     * @param channelType the channel type. ie messaging
-     * @param channelId the channel id. ie 123
-     */
-    @Deprecated(
-        message = "Use getChannelClient() instead",
-        replaceWith = ReplaceWith("this.getChannelClient")
-    )
-    public fun channel(channelType: String, channelId: String): ChannelController {
-        return ChannelControllerImpl(channelType, channelId, this)
-    }
-
-    /***
-     * Returns a [ChannelController] for given cid
-     *
-     * @param cid the full channel id. ie messaging:123
-     */
-    @Deprecated(
-        message = "Use getChannelClient() instead",
-        replaceWith = ReplaceWith("this.getChannelClient")
-    )
-    public fun channel(cid: String): ChannelController {
-        val type = cid.split(":")[0]
-        val id = cid.split(":")[1]
-        return channel(type, id)
-    }
-
-    /***
      * Returns a [ChannelClient] for given type and id
      *
      * @param channelType the channel type. ie messaging
      * @param channelId the channel id. ie 123
      */
-    public fun getChannelClient(channelType: String, channelId: String): ChannelClient {
+    public fun channel(channelType: String, channelId: String): ChannelClient {
         return ChannelClient(channelType, channelId, this)
     }
 
@@ -759,10 +728,10 @@ public class ChatClient internal constructor(
      *
      * @param cid the full channel id. ie messaging:123
      */
-    public fun getChannelClient(cid: String): ChannelClient {
+    public fun channel(cid: String): ChannelClient {
         val type = cid.split(":")[0]
         val id = cid.split(":")[1]
-        return getChannelClient(type, id)
+        return channel(type, id)
     }
 
     public fun createChannel(
